api_ma/requirements.txt
fastapi
uvicorn
peft
sentence-transformers
faiss-cpu
datasets
spacy
spacy-lookups-data
pyspellchecker
sacremoses
transformers
torch
pandas
requests
numpy
sentencepiece
psutil
urllib3==1.26.6
bs4
lxml
fastapi-limiter
python-dotenv
cryptography
gunicorn
watchdog
faiss-cpu